+++ packages/index.js
@@ -65,7 +65,7 @@ function install (Vue, opts = {}) {
 }
 
 export default {
-    version: '1.6.8',
+    version: '1.6.9',
     install,
     LCell,
     LDateRange,